+++ src/app/components/file-browser/storage-location-per-provider-table.js
@@ -8,10 +8,14 @@
 
 import Component from '@ember/component';
 import { reads } from '@ember/object/computed';
+import { computed } from '@ember/object';
+import { inject as service } from '@ember/service';
 
 export default Component.extend({
   tagName: 'tbody',
   classNames: ['storage-location-per-provider-table'],
+
+  errorExtractor: service(),
 
   /**
    * @virtual
@@ -23,4 +27,12 @@ export default Component.extend({
    * @type {ComputedProperty<Models.Provider>}
    */
   oneprovider: reads('locations.firstObject.provider'),
+
+  /**
+   * @type {ComputedProperty<string>}
+   */
+  errorMessage: computed('locations.firstObject.error', function errorMessage() {
+    return this.errorExtractor.getMessage(this.locations.firstObject.error)?.message ||
+      this.locations.firstObject.error.description;
+  }),
 });