+++ src/app/locales/en/components/file-info-modal.js
@@ -1,3 +1,5 @@
+const fileLinkAuthentication = 'Requires user authentication and the necccessary privileges';
+
 export default {
   header: '{{type}} details',
   headerDefault: 'Details',
@@ -12,7 +14,20 @@ export default {
   modificationTime: 'Modified at',
   owner: 'Owner',
   size: 'Size',
-  guiUrl: '{{type}} URL',
+  fileLinkLabel: {
+    show: 'Show item link',
+    download: 'Download item link',
+  },
+  fileLinkTip: {
+    show: {
+      file: `<p><strong>Show item link</strong></p><p>The URL opens a file browser view with the file selected. ${fileLinkAuthentication} for the file.</p>`,
+      dir: `<p><strong>Show item link</strong></p><p>The URL opens a file browser view with the directory selected. ${fileLinkAuthentication} for the directory.</p>`,
+    },
+    download: {
+      file: `<p><strong>Download item link</strong></p><p>The URL opens a file browser view and asks for the file download. ${fileLinkAuthentication} for the file.</p><p>If you are looking for a URL to download the file without web browser usage, check out the API tab.</p>`,
+      dir: `<p><strong>Download item link</strong></p><p>The URL opens a file browser view and asks for the directory and its contents download, packaged in the tar format. ${fileLinkAuthentication} for the directory.</p><p>If you are looking for a URL to download the directory without web browser usage, check out the API tab.</p>`,
+    },
+  },
   fileType: {
     file: 'file',
     dir: 'directory',